requirements.txt
aiofiles==22.1.0
aiosmtplib==1.1.7
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
blinker==1.5
boto3==1.26.3
botocore==1.29.3
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
databases==0.6.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
fastapi-mail==1.2.0
greenlet==1.1.3.post0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
install==1.3.5
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.2
ormar==0.11.3
passlib==1.7.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
